refactor: separate FragmentAdd, FragmentDelete, and FragmentSetProperties into individual files
